requirements.txt
APScheduler==3.6.1
certifi==2019.6.16
chardet==3.0.4
convertapi==1.1.0
dateparser==0.7.1
Django==2.2.3
django-cors-headers==3.0.2
django-pandas==0.6.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
idna==2.8
lxml==4.4.0
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.1
regex==2019.6.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.3
